added normalization utility

+++ src/main.rs
@@ -10,7 +10,7 @@ use std::fs::File;
 use std::env::args;
 use std::path::Path;
 
-fn read_csv(path: &str, has_headers: bool) -> Result<Array2<f32>, ReadError> {
+fn read_csv(path: &str, has_headers: bool) -> Result<Array2<f64>, ReadError> {
     let file = match File::open(path) {
         Ok(file) => file,
         Err(err) => panic!("Cannot open file \"{path}\": {err}"),
@@ -18,16 +18,16 @@ fn read_csv(path: &str, has_headers: bool) -> Result<Array2<f32>, ReadError> {
     let mut reader = ReaderBuilder::new()
         .has_headers(has_headers)
         .from_reader(file);
-    let array_read: Array2<f32> = reader.deserialize_array2_dynamic()?;
+    let array_read: Array2<f64> = reader.deserialize_array2_dynamic()?;
 
     Ok(array_read)
 }
 
-fn linear_estimate(estimators: &Array2<f32>, weights: &Array1<f32>) -> Array1<f32> {
+fn linear_estimate(estimators: &Array2<f64>, weights: &Array1<f64>) -> Array1<f64> {
     // Calculates linear estimates across all observations.
 
     let nrows = estimators.nrows();
-    let mut estimates: Array1<f32> = Array::zeros((nrows,));
+    let mut estimates: Array1<f64> = Array::zeros((nrows,));
 
     for i in 0..nrows {
         estimates[i] = estimators.index_axis(Axis(0), i).dot(weights);
@@ -37,10 +37,10 @@ fn linear_estimate(estimators: &Array2<f32>, weights: &Array1<f32>) -> Array1<f3
 }
 
 fn mean_sqsum_loss_gradient(
-    estimators: &Array2<f32>,
-    observations: &Array1<f32>,
-    weights: &Array1<f32>,
-) -> Array1<f32> {
+    estimators: &Array2<f64>,
+    observations: &Array1<f64>,
+    weights: &Array1<f64>,
+) -> Array1<f64> {
     // loss = sum ( (x_i[] * theta_i[] - y_i)^2 ) where i is the number of observations
     // grad(loss(...)) = 1/m * sum ( 2 * (sum_col(x_i[] * theta_i[]) - y_i) * x_i[] )
 
@@ -54,12 +54,12 @@ fn mean_sqsum_loss_gradient(
         tmp_weights[j] = estimators.index_axis(Axis(1), j).dot(&diff);
     }
 
-    tmp_weights / (observations.len() as f32)
+    tmp_weights / (observations.len() as f64)
 }
 
-fn gradient_descent2(data: &Array2<f32>, lr: f32, min_change: f32) -> Array1<f32> {
+fn gradient_descent2(data: &Array2<f64>, lr: f64, min_change: f64) -> Array1<f64> {
     const ITER_MAX: usize = 10000;
-    let mut weights: Array1<f32> = array![0., 0.];
+    let mut weights: Array1<f64> = array![0., 0.];
     let estimators = stack!(
         // Prepend 1 for x_0
         Axis(1),
@@ -72,8 +72,8 @@ fn gradient_descent2(data: &Array2<f32>, lr: f32, min_change: f32) -> Array1<f32
     println!("Learning rate: {lr}");
     for _i in 0..ITER_MAX {
         let change = lr * mean_sqsum_loss_gradient(&estimators, &price, &weights);
-        let change_abs: Array1<f32> = change.iter().map(|x| x.abs()).collect();
-        let max = change_abs.iter().fold(f32::NEG_INFINITY, |a, &b| a.max(b));
+        let change_abs: Array1<f64> = change.iter().map(|x| x.abs()).collect();
+        let max = change_abs.iter().fold(f64::NEG_INFINITY, |a, &b| a.max(b));
 
         if max < min_change {
             break;
@@ -108,10 +108,10 @@ fn main() {
     let learn_rate = args()
         .nth(2)
         .expect("no learning rate provided")
-        .parse::<f32>()
+        .parse::<f64>()
         .expect("cannot parse learning rate");
     let min_change = match args().nth(3) {
-        Some(val) => val.parse::<f32>().expect("cannot parse min_change"),
+        Some(val) => val.parse::<f64>().expect("cannot parse min_change"),
         None => 1e-5,
     };
